requirements.txt
annotated-types==0.7.0
anyio==4.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.2
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
comtypes==1.4.6
crispy-bootstrap4==2025.6
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.2.3
django-crispy-forms==2.4
django-jalali-date==2.0.0
executing==2.1.0
fastapi==0.115.14
fastjsonschema==2.20.0
filelock==3.18.0
fqdn==1.5.1
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.7
ipykernel==6.29.5
ipython==8.27.0
isoduration==20.11.0
jalali_core==1.0.0
jdatetime==5.2.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
oauth2client==4.1.3
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pillow==12.0.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==6.30.2
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyDrive==1.3.1
Pygments==2.18.0
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==2.0.7
pyttsx3 @ file:///E:/instalations/pyttsx3-2.91-py3-none-any.whl#sha256=8f19217c4d65a46870e0b82f2e7eb1570ec2b2ca80517fbeee095e54f76f73e2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
rsa==4.9.1
Send2Trash==1.8.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition @ file:///E:/instalations/SpeechRecognition-3.10.4-py2.py3-none-any.whl#sha256=723b8155692a8ed11a30013f15f89a3e57c5dc8bc73c8cb024bf9bd14c21fba5
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
terminado==0.18.1
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.35.0
virtualenv==20.31.2
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
